@@ -72,7 +72,8 @@
     "babel-plugin-source-map-support": "^2.0.1",
     "commander": "^2.19.0",
     "cpx": "^1.5.0",
-    "electron": "^4.1.1",
+    "electron": "^4.2.4",
+    "electron-devtools-installer": "^2.2.4",
     "electron-installer-dmg": "^2.0.0",
     "electron-packager": "^13.1.1",
     "enzyme": "^3.9.0",
@@ -86,10 +87,11 @@
     "flow-typed": "^2.5.1",
     "import-js": "^3.2.0",
     "jest": "^24.5.0",
-    "livereload": "^0.7.0",
+    "livereload": "^0.8.0",
     "node-sass-chokidar": "^1.3.3",
     "npm-run-all": "^4.1.3",
     "prettier": "^1.16.4",
+    "prettier-eslint": "^8.8.2",
     "spectron": "^5.0.0",
     "whatwg-fetch": "^3.0.0"
   },
@@ -97,7 +99,6 @@
     "base64-js": "^1.3.0",
     "classnames": "^2.2.6",
     "d3": "^5.9.2",
-    "electron-devtools-installer": "^2.2.4",
     "electron-dl": "^1.13.0",
     "electron-is-dev": "^1.0.1",
     "fs-extra": "^7.0.0",
