@@ -1,13 +1,12 @@
 /* @flow */
 
-import {connect} from "react-redux"
+import {useSelector} from "react-redux"
 import React from "react"
 
 import {Md5Panel} from "./Md5Panel"
 import type {Search} from "../../state/searches/types"
-import type {State} from "../../state/types"
 import {buildLogDetail, getIsGoingBack} from "../../state/selectors/logDetails"
-import {getSearchesByTag} from "../../state/searches/selector"
+import {getLogDetailSearches} from "../../state/searches/selector"
 import ConnPanel from "./ConnPanel"
 import FieldsPanel from "./FieldsPanel"
 import Log from "../../models/Log"
@@ -15,45 +14,27 @@ import NavAnimation from "./NavAnimation"
 import NoSelection from "./NoSelection"
 import UidPanel from "./UidPanel"
 
-type Props = {|
-  log: ?Log,
-  isGoingBack: boolean,
-  searches: Search[]
-|}
-
 export type PanelProps = {|
   searches: Search[],
   log: Log
 |}
 
-export default class LogDetails extends React.Component<Props> {
-  render() {
-    const {log, searches, isGoingBack} = this.props
-    if (!log) return <NoSelection />
+export default function LogDetails() {
+  let log = useSelector(buildLogDetail)
+  let isGoingBack = useSelector(getIsGoingBack)
+  let searches = useSelector(getLogDetailSearches)
+  let panelProps = {log, searches}
 
-    const panelProps = {log, searches}
+  if (!log) return <NoSelection />
 
-    return (
-      <NavAnimation log={log} prev={isGoingBack}>
-        <div className="log-detail">
-          <FieldsPanel {...panelProps} />
-          <UidPanel {...panelProps} />
-          <ConnPanel {...panelProps} />
-          <Md5Panel {...panelProps} />
-        </div>
-      </NavAnimation>
-    )
-  }
+  return (
+    <NavAnimation log={log} prev={isGoingBack}>
+      <div className="log-detail">
+        <FieldsPanel {...panelProps} />
+        <UidPanel {...panelProps} />
+        <ConnPanel {...panelProps} />
+        <Md5Panel {...panelProps} />
+      </div>
+    </NavAnimation>
+  )
 }
-
-function stateToProps(state: State) {
-  return {
-    log: buildLogDetail(state),
-    isGoingBack: getIsGoingBack(state),
-    searches: getSearchesByTag(state, "detail")
-  }
-}
-
-export const XLogDetails = connect<Props, {||}, _, _, _, _>(stateToProps)(
-  LogDetails
-)
