@@ -8,33 +8,18 @@ type Props = {
   onDrag: Function,
   children: *,
   width: number,
-  className: string,
-  onMouseEnter: Function,
-  onMouseLeave: Function
+  className: string
 }
 
 export default class Pane extends React.Component<Props> {
   render() {
     if (!this.props.isOpen) return null
 
-    const {
-      position,
-      onDrag,
-      children,
-      width,
-      className,
-      onMouseEnter,
-      onMouseLeave
-    } = this.props
+    const {position, onDrag, children, width, className} = this.props
     const anchorPos = position === "left" ? "right" : "left"
 
     return (
-      <aside
-        className={`pane pane-${position} ${className}`}
-        style={{width}}
-        onMouseEnter={onMouseEnter}
-        onMouseLeave={onMouseLeave}
-      >
+      <aside className={`pane pane-${position} ${className}`} style={{width}}>
         {children}
         <DragAnchor onDrag={onDrag} position={anchorPos} />
       </aside>
