@@ -6,8 +6,6 @@ import classNames from "classnames"
 
 import type {DispatchProps} from "../state/types"
 import type {Space} from "../lib/Space"
-import {XLogDetails} from "./LogDetails"
-import {XRightPaneCollapser} from "./RightPaneCollapser"
 import {XRightPaneExpander} from "./RightPaneExpander"
 import {
   backLogDetail,
@@ -21,6 +19,7 @@ import {open} from "../lib/System"
 import Back from "../icons/back-arrow.svg"
 import Forward from "../icons/forward-arrow.svg"
 import Log from "../models/Log"
+import LogDetails from "./LogDetails"
 import Pane, {
   PaneHeader,
   PaneTitle,
@@ -29,6 +28,7 @@ import Pane, {
   Center,
   PaneBody
 } from "./Pane"
+import RightPaneCollapser from "./RightPaneCollapser"
 import Star from "../icons/star-sm.svg"
 import dispatchToProps from "../lib/dispatchToProps"
 import * as logDetails from "../state/selectors/logDetails"
@@ -92,12 +92,6 @@ export default class RightPane extends React.Component<Props, S> {
         position="right"
         width={width}
         className="log-detail-pane"
-        onMouseEnter={() => {
-          this.setState({showCollapse: true})
-        }}
-        onMouseLeave={() => {
-          this.setState({showCollapse: false})
-        }}
       >
         {currentLog && (
           <PaneHeader>
@@ -143,9 +137,9 @@ export default class RightPane extends React.Component<Props, S> {
           </PaneHeader>
         )}
         <PaneBody>
-          <XLogDetails />
+          <LogDetails />
         </PaneBody>
-        <XRightPaneCollapser show={this.state.showCollapse} />
+        <RightPaneCollapser />
       </Pane>
     )
   }
